refactor: split 분리
