+++ src/test/java/stringAddCalculate/StringAddCalculatorTest.java
@@ -8,40 +8,40 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 public class StringAddCalculatorTest {
     @Test
     public void splitAndSum_null_또는_빈문자() {
-        int result = StringAddCalculator.splitAndSum(null);
+        int result = StringAddCalculator.calculate(null);
         assertThat(result).isEqualTo(0);
 
-        result = StringAddCalculator.splitAndSum("");
+        result = StringAddCalculator.calculate("");
         assertThat(result).isEqualTo(0);
     }
 
     @Test
     public void splitAndSum_숫자하나() throws Exception {
-        int result = StringAddCalculator.splitAndSum("1");
+        int result = StringAddCalculator.calculate("1");
         assertThat(result).isEqualTo(1);
     }
 
     @Test
     public void splitAndSum_쉼표구분자() throws Exception {
-        int result = StringAddCalculator.splitAndSum("1,2");
+        int result = StringAddCalculator.calculate("1,2");
         assertThat(result).isEqualTo(3);
     }
 
     @Test
     public void splitAndSum_쉼표_또는_콜론_구분자() throws Exception {
-        int result = StringAddCalculator.splitAndSum("1,2:3");
+        int result = StringAddCalculator.calculate("1,2:3");
         assertThat(result).isEqualTo(6);
     }
 
     @Test
     public void splitAndSum_custom_구분자() throws Exception {
-        int result = StringAddCalculator.splitAndSum("//;\n1;2;3");
+        int result = StringAddCalculator.calculate("//;\n1;2;3");
         assertThat(result).isEqualTo(6);
     }
 
     @Test
     public void splitAndSum_negative() throws Exception {
-        assertThatThrownBy(() -> StringAddCalculator.splitAndSum("-1,2,3"))
+        assertThatThrownBy(() -> StringAddCalculator.calculate("-1,2,3"))
                 .isInstanceOf(RuntimeException.class);
     }
 }
